test(agent-manager): avoid tsx import in annotation test
